pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "synthatrial"
version = "0.2.0"
description = "In Silico Pharmacogenomics Platform with Agentic AI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SynthaTrial Team"},
]
keywords = ["pharmacogenomics", "ai", "drug-discovery", "personalized-medicine"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "rdkit>=2023.9.1",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "pinecone>=5.0.0",
    "langchain-google-genai>=0.1.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "pytest>=7.0.0",
    "hypothesis>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.6.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "bandit>=1.7.5",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-mock>=3.12.0",
    "pytest-benchmark>=4.0.0",
    "pytest-html>=4.1.0",
    "pytest-json-report>=1.5.0",
    "pytest-timeout>=2.2.0",
    "pytest-asyncio>=0.23.0",
    "coverage>=7.4.0",
]

[project.urls]
Homepage = "https://github.com/synthatrial/synthatrial"
Documentation = "https://github.com/synthatrial/synthatrial/docs"
Repository = "https://github.com/synthatrial/synthatrial"
"Bug Tracker" = "https://github.com/synthatrial/synthatrial/issues"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | data
  | logs
  | \.hypothesis
  | \.pytest_cache
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["data/*", "logs/*", ".hypothesis/*", ".pytest_cache/*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "data/",
    "logs/",
    ".hypothesis/",
    ".pytest_cache/",
    "tests/",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "property: Property-based test using hypothesis",
    "ssl_generation: SSL certificate generation tests",
    "ssl_expiration: SSL certificate expiration tests",
    "ssl_validation: SSL certificate validation tests",
    "ssl_error_handling: SSL error handling tests",
    "data_download: Data download tests",
    "container_startup: Container startup validation tests",
    "progress_tracking: Download progress tracking tests",
    "data_integrity: Data integrity validation tests",
    "chembl_validation: ChEMBL database validation tests",
    "integration: Integration tests",
    "unit: Unit tests",
    "slow: Slow running tests",
    "fast: Fast running tests",
    "dev: Development environment tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/.*",
    "data/*",
    "logs/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "data", "logs", ".hypothesis", ".pytest_cache"]
skips = ["B101", "B601"]
